Final commit push. Ready for review:
Today i added the video element with the sources it needed in order to play in the actual rendered page. Next, i added a track element to be able to play captions with it, then i added css styles to affect the video itself on the rendered page. Lastly i added a media query to affect the movement of the video box and i affected the caption cues with a background color, text color and a font style.